tutorial/requirements.txt
djangorestframework==3.11.2
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
drf-spectacular==0.26.3
Django==3.1.8
Pygments==2.7.4
Markdown==3.2.1
coreapi==2.3.3
psycopg2-binary==2.9.3
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.0.1
PyYAML==5.4
django-extensions==3.1.5
ipython
